`static_cast<unsigned int>(re.flags())`

+++ tests/std/tests/GH_005553_regex_character_translation/test.cpp
@@ -56,7 +56,7 @@ template <class Rx>
 void check_match(const wstring& subject, const wstring& pattern, const Rx& re, match_flag_type flags, bool matches) {
     if (regex_match(subject, re, flags) != matches) {
         wprintf(LR"(Expected regex_match("%s", regex("%s", 0x%X)) to be %s.)", subject.c_str(), pattern.c_str(),
-            re.flags(), matches ? L"true" : L"false");
+            static_cast<unsigned int>(re.flags()), matches ? L"true" : L"false");
         g_regexTester.fail_regex();
     }
 }
